Added some defensive programming to validation function.
Added some defensive programming to validation function. Making sure that return type of the function is as expected, which is either `null` or `string`. I did this because it make sense to have validation functions in a curry/factory pattern such as:

```

const numberIsInRangeOf = (min, max) => (value) => value > min && value < max` ? null : Number should be within ${min} and ${max}`

```

But then, it's easy to forget to invoke a function, and instead of passing validation function, pass a factory:

```
// Good - passing validation function
createRule([numberIsInRangeOf(100, 200)])
// Bad - forgot to pass min/max, in this case function returns `function` instead of `?String`
createRule([numberIsInRangeOf])
```